+++ firmware/Core/Inc/stm32/stm32_internal_flash_drivers.h
@@ -1,5 +1,5 @@
-#ifndef __INCLUDE_GUARD__STM32_INTERNAL_FLASH_DRIVERS_H__
-#define __INCLUDE_GUARD__STM32_INTERNAL_FLASH_DRIVERS_H__
+#ifndef INCLUDE_GUARD__STM32_INTERNAL_FLASH_DRIVERS_H__
+#define INCLUDE_GUARD__STM32_INTERNAL_FLASH_DRIVERS_H__
 
 #include <stdint.h>
 
@@ -37,4 +37,4 @@ uint8_t STM32_internal_flash_set_active_flash_bank(uint8_t wanted_active_flash_b
 
 uint8_t STM32_internal_flash_get_active_flash_bank();
 
-#endif // __INCLUDE_GUARD__STM32_INTERNAL_FLASH_DRIVERS_H__
+#endif // INCLUDE_GUARD__STM32_INTERNAL_FLASH_DRIVERS_H__
