+++ firmware/Core/Inc/telecommand_exec/telecommand_definitions.h
@@ -3,7 +3,7 @@
 #define INCLUDE_GUARD__TELECOMMAND_DEFINITIONS_H
 
 #include <stdint.h>
-#include "telecommands/telecommand_types.h"
+#include "telecommand_exec/telecommand_types.h"
 
 extern const TCMD_TelecommandDefinition_t TCMD_telecommand_definitions[];
 extern const int16_t TCMD_NUM_TELECOMMANDS;
