+++ firmware/Core/Inc/telecommands/freertos_telecommand_defs.h
@@ -2,7 +2,7 @@
 #ifndef INCLUDE_GUARD__FREERTOS_TELECOMMAND_DEFINITIONS_H
 #define INCLUDE_GUARD__FREERTOS_TELECOMMAND_DEFINITIONS_H
 
-#include "telecommands/telecommand_types.h"
+#include "telecommand_exec/telecommand_types.h"
 
 #include <stdint.h>
