+++ firmware/Core/Inc/telecommands/stm32_internal_flash_telecommand_defs.h
@@ -1,7 +1,7 @@
 #ifndef INCLUDE_GUARD_STM32_INTERNAL_FLASH_TELECOMMAND_DEFS_H
 #define INCLUDE_GUARD_STM32_INTERNAL_FLASH_TELECOMMAND_DEFS_H
 
-#include "telecommands/telecommand_types.h"
+#include "telecommand_exec/telecommand_types.h"
 
 uint8_t TCMDEXEC_stm32_internal_flash_write(const char *args_str, TCMD_TelecommandChannel_enum_t tcmd_channel,
                                             char *response_output_buf, uint16_t response_output_buf_len);
