+++ firmware/Core/Inc/telecommands/system_telecommand_defs.h
@@ -2,7 +2,7 @@
 #define INCLUDE_GUARD__SYSTEM_TELECOMMAND_DEFINITIONS_H
 
 #include <stdint.h>
-#include "telecommands/telecommand_types.h"
+#include "telecommand_exec/telecommand_types.h"
 
 uint8_t TCMDEXEC_hello_world(const char *args_str, TCMD_TelecommandChannel_enum_t tcmd_channel,
     char *response_output_buf, uint16_t response_output_buf_len);
