+++ firmware/Core/Inc/telecommands/temperature_sensor_telecommand_defs.h
@@ -1,7 +1,7 @@
 #ifndef INCLUDE_GUARD__TEMPERATURE_SENSOR_TELECOMMAND_DEFS_H__
 #define INCLUDE_GUARD__TEMPERATURE_SENSOR_TELECOMMAND_DEFS_H__
 
-#include "telecommands/telecommand_definitions.h"
+#include "telecommand_exec/telecommand_definitions.h"
 
 
 uint8_t TCMDEXEC_obc_read_temperature(const char *args_str, TCMD_TelecommandChannel_enum_t tcmd_channel,
