+++ firmware/Core/Inc/telecommands/testing_telecommand_defs.h
@@ -2,7 +2,7 @@
 #ifndef INCLUDE_GUARD__TESTING_TELECOMMAND_DEFINITIONS_H
 #define INCLUDE_GUARD__TESTING_TELECOMMAND_DEFINITIONS_H
 
-#include "telecommands/telecommand_types.h"
+#include "telecommand_exec/telecommand_types.h"
 #include <stdint.h>
 
 uint8_t TCMDEXEC_echo_back_args(const char *args_str, TCMD_TelecommandChannel_enum_t tcmd_channel,
