+++ firmware/Core/Inc/telecommands/timekeeping_telecommand_defs.h
@@ -3,7 +3,7 @@
 #define INCLUDE_GUARD__TIMEKEEPING_TELECOMMAND_DEFS_H__
 
 #include <stdint.h>
-#include "telecommands/telecommand_definitions.h"
+#include "telecommand_exec/telecommand_definitions.h"
 
 uint8_t TCMDEXEC_get_system_time(const char *args_str, TCMD_TelecommandChannel_enum_t tcmd_channel,
                         char *response_output_buf, uint16_t response_output_buf_len);
