+++ firmware/Core/Inc/telecommands/uart_telecommand_defs.h
@@ -1,7 +1,7 @@
 #ifndef INCLUDE_GUARD__UART_TELECOMMAND_DEFS_H__
 #define INCLUDE_GUARD__UART_TELECOMMAND_DEFS_H__
 
-#include "telecommands/telecommand_definitions.h"
+#include "telecommand_exec/telecommand_definitions.h"
 
 uint8_t TCMDEXEC_uart_send_hex_get_response_hex(
     const char *args_str, TCMD_TelecommandChannel_enum_t tcmd_channel,
