+++ firmware/Core/Src/rtos_tasks/rtos_tasks.c
@@ -2,8 +2,8 @@
 #include "main.h"
 #include "rtos_tasks/rtos_task_helpers.h"
 #include "rtos_tasks/rtos_tasks.h"
-#include "telecommands/telecommand_parser.h"
-#include "telecommands/telecommand_executor.h"
+#include "telecommand_exec/telecommand_parser.h"
+#include "telecommand_exec/telecommand_executor.h"
 #include "debug_tools/debug_uart.h"
 #include "timekeeping/timekeeping.h"
 #include "uart_handler/uart_handler.h"
