+++ firmware/Core/Src/telecommand_exec/telecommand_args_helpers.c
@@ -1,5 +1,5 @@
 #include "transforms/arrays.h"
-#include "telecommands/telecommand_args_helpers.h"
+#include "telecommand_exec/telecommand_args_helpers.h"
 
 #include <string.h>
 #include <stdint.h>
