+++ firmware/Core/Src/telecommand_exec/telecommand_parser.c
@@ -1,6 +1,6 @@
-#include "telecommands/telecommand_definitions.h"
-#include "telecommands/telecommand_parser.h"
-#include "telecommands/telecommand_args_helpers.h"
+#include "telecommand_exec/telecommand_definitions.h"
+#include "telecommand_exec/telecommand_parser.h"
+#include "telecommand_exec/telecommand_args_helpers.h"
 #include "debug_tools/debug_uart.h"
 #include "transforms/arrays.h"
 #include "crypto/sha256.h"
