+++ firmware/Core/Src/telecommands/agenda_telecommand_defs.c
@@ -1,6 +1,6 @@
-#include "telecommands/telecommand_definitions.h"
-#include "telecommands/telecommand_args_helpers.h"
-#include "telecommands/telecommand_executor.h"
+#include "telecommand_exec/telecommand_definitions.h"
+#include "telecommand_exec/telecommand_args_helpers.h"
+#include "telecommand_exec/telecommand_executor.h"
 #include "debug_tools/debug_uart.h"
 #include "telecommands/agenda_telecommands_defs.h"
 #include "log/log.h"
