+++ firmware/Core/Src/telecommands/freertos_telecommand_defs.c
@@ -1,6 +1,6 @@
 
-#include "telecommands/telecommand_definitions.h"
-#include "telecommands/telecommand_args_helpers.h"
+#include "telecommand_exec/telecommand_definitions.h"
+#include "telecommand_exec/telecommand_args_helpers.h"
 #include "transforms/arrays.h"
 #include "unit_tests/unit_test_executor.h"
 #include "timekeeping/timekeeping.h"
