+++ firmware/Core/Src/telecommands/mpi_telecommand_defs.c
@@ -1,4 +1,4 @@
-#include "telecommands/telecommand_args_helpers.h"
+#include "telecommand_exec/telecommand_args_helpers.h"
 #include "telecommands/mpi_telecommand_defs.h"
 #include "mpi/mpi_command_handling.h"
 #include "transforms/arrays.h"
